refactor(drivers/postgres): Move and rename "schema/postgres" to "drivers/postgres/schema"
BREAKING CHANGE: This commit moves the "schema/postgres" package to "drivers/postgres/schema". This change aligns the package location and name with its purpose of providing utilities for managing PostgreSQL schemas in a multi-tenant application. The relocation and renaming improve code organization and clarity. This is a breaking change and requires updates in the code where this package is imported.