feat(scopes): Update `WithTenantSchema` to support pointer to array/slice of structs
This commit modifies the `scopes` package to improve the logic for retrieving table names. The previous `tableNameFromReflectValue` was renamed to  `tableNameFromInterface`. Additionally, the `tableNameFromReflectValue` is now responsible for handling reflect values of arrays, and slices. These changes enhance the reliability and flexibility of table name retrieval in the `scopes` package.