chore: Add release script
This commit adds a new release script and updates the Makefile to include a new target for creating a new release. The release script is used to release the project in all directories that contain a .releaserc.json file. It installs dependencies and then runs the release command for each directory.